http: move request helper into utils
